fix: resolve all clippy linting errors for CI
- Allow deprecated Nonce::from_slice calls with #[allow(deprecated)]
- Remove unnecessary type casts in format.rs
- Use div_ceil() method instead of manual ceiling division
- Remove needless as_bytes() call in local_entry.rs
- Use collapsible str::replace with character array
- Fix empty lines after doc comments in examples
- Add explicit type annotation for div_ceil ambiguity

All 43 tests passing, clippy clean with -D warnings

+++ src/archive/format.rs
@@ -230,7 +230,7 @@ impl FileHeader {
     pub fn validate_version(&self) -> Result<()> {
         if self.version_major > FORMAT_VERSION_MAJOR {
             return Err(EngramError::UnsupportedVersion(
-                (self.version_major as u16) << 8 | self.version_minor as u16,
+                self.version_major << 8 | self.version_minor,
             ));
         }
         Ok(())
